Extract WTA app contracts
Move shared App events and payload models behind a neutral contract boundary, and separate WT event publishing from App state.


Copilot-Session: ad8b225c-0d7f-4eed-8d87-5965030cf647
