+++ tools/wta/src/app_contracts/agent.rs
@@ -1,0 +1,8 @@
+use crate::agent_source::AgentSource;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct AvailableAgent {
+    pub id: String,
+    pub display_name: String,
+    pub source: AgentSource,
+}
