+++ tools/wta/src/app_contracts/diagnostics.rs
@@ -1,0 +1,12 @@
+#[derive(Debug, Clone)]
+pub enum DebugDir {
+    Sent,
+    Received,
+}
+
+#[derive(Debug, Clone)]
+pub struct DebugMessage {
+    pub timestamp: f64,
+    pub direction: DebugDir,
+    pub content: String,
+}
