+++ tools/wta/src/app_contracts/model.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct AcpModelInfo {
+    pub id: String,
+    pub name: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+}
