+++ tools/wta/src/main.rs
@@ -8,6 +8,7 @@ mod agent_registry;
 mod agent_source;
 mod agent_sessions;
 mod app;
+mod app_contracts;
 mod clipboard_image;
 mod command_recall;
 mod commands;
@@ -41,6 +42,7 @@ mod ui_trace;
 mod win32;
 mod wsl;
 mod wsl_acp;
+mod wt_protocol_events;
 
 use agent_client_protocol as acp;
 use anyhow::{bail, Context, Result};
